+++ tests/test_profile.py
@@ -19,6 +19,21 @@ def simple_profile_data() -> DictFromJSON:
     }
 
 
+@pytest.fixture
+def complex_profile_data() -> DictFromJSON:
+    """Profile data with all implemented fields populated.
+
+    Mocks dict returned by `Spond.get_group()['members'][n][profile]`.
+    """
+    return {
+        "id": "P2",
+        "firstName": "Salma",
+        "lastName": "Hayek",
+        "email": "s_hayek@example.com",
+        "phoneNumber": "+123456790",
+    }
+
+
 def test_from_dict(simple_profile_data: DictFromJSON) -> None:
     """Test that Profile is created from the simplest possible data dict."""
     # arrange
@@ -27,3 +42,14 @@ def test_from_dict(simple_profile_data: DictFromJSON) -> None:
     # assert
     assert my_profile.uid == "P1"
     assert str(my_profile) == "Profile(uid='P1', full_name='Morgan Freeman', …)"
+
+
+def test_from_dict_additional_fields(complex_profile_data: DictFromJSON) -> None:
+    """Test that Profile can be created with additional supported data."""
+    # arrange
+    # act
+    my_profile = Profile(**complex_profile_data)
+    # assert
+    # - optional:
+    assert my_profile.email == "s_hayek@example.com"
+    assert my_profile.phone_number == "+123456790"